Send AcknowledgeBlockChange packet after interaction
