+++ crates/core/src/player.rs
@@ -10,9 +10,8 @@ use mchprs_blocks::{BlockDirection, BlockFacing, BlockPos};
 use mchprs_network::packets::clientbound::*;
 use mchprs_network::packets::{PacketEncoder, SlotData};
 use mchprs_network::{PlayerConn, PlayerPacketSender};
-use mchprs_text::{ColorCode, TextColor, TextComponent, TextComponentBuilder};
+use mchprs_text::{ColorCode, TextComponent, TextComponentBuilder};
 use serde::{Deserialize, Serialize};
-use serde_json::json;
 use std::fmt;
 use std::fs::{self, OpenOptions};
 use std::io::{Cursor, Write};
